+++ encodesrv/log.py
@@ -4,10 +4,10 @@ Author: Robert Walker <robert.walker@ystv.co.uk> 2015
 """
 
 import logging.handlers
-import bots.slack
-import bots.irc
+from .bots import slack
+from .bots import irc
 
-from config import Config
+from .config import Config
 
 LOG_FILENAME= "/tmp/encodesrv.log"
 LOG_FORMAT = '%(asctime)s:%(levelname)s:%(message)s'
@@ -35,13 +35,13 @@ def setup_logging(encodesrv_daemon):
     
     # Slack bot logging
     if Config['slack']['enabled']:
-        slackhandler = bots.slack.Encode_slack(encodesrv_daemon, **Config['slack'])
+        slackhandler = slack.Encode_slack(encodesrv_daemon, **Config['slack'])
         slackhandler.setLevel(logging.INFO)
         logging.getLogger().addHandler(slackhandler)
 
     # IRC bot logging
     if Config['irc']['enabled']:
-        irchandler = bots.irc.Encode_irc(encodesrv_daemon, **Config['irc'])
+        irchandler = irc.Encode_irc(encodesrv_daemon, **Config['irc'])
         
         while not irchandler.is_joined():
             pass
